requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
gunicorn==20.1.0
nltk==3.8.1
networkx==2.8.8
typing-extensions==4.5.0
assemblyai==0.17.0
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
pyOpenSSL==23.2.0
cryptography==41.0.7
pytz==2023.3
openai==1.1.0
pandas==2.0.3
